Upgrade styled-system to 5.0.18. Extract ThemeProvider into own module, to prevent circular dependency. Upgrade tsconfig files to use same build dir as rollup.
