+++ packages/elements/package.json
@@ -35,10 +35,10 @@
     "@fortawesome/react-fontawesome": ">=0.1.4",
     "@types/react": ">=16.8.19",
     "@types/react-dom": ">=16.8.4",
-    "@types/styled-system": ">=4.1.0",
+    "@types/styled-system": ">=5.0.1",
     "react": ">=16.8.6",
     "react-dom": ">=16.8.6",
-    "styled-system": ">=4.1.0"
+    "styled-system": ">=5.0.18"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-svg-core": "1.2.18",
@@ -47,10 +47,10 @@
     "@types/jest": "^23.1.5",
     "@types/react": "^16.8.19",
     "@types/react-dom": "^16.8.4",
-    "@types/styled-system": "^4.1.0",
+    "@types/styled-system": "^5.0.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "styled-system": "^4.1.0"
+    "styled-system": "^5.0.18"
   },
   "files": [
     "dist"
