+++ packages/forms/package.json
@@ -44,12 +44,12 @@
     "@types/lodash": ">=4.14.135",
     "@types/react": ">=16.8.19",
     "@types/react-dom": ">=16.8.4",
-    "@types/styled-system": ">=4.1.0",
+    "@types/styled-system": ">=5.0.1",
     "lodash": ">=4.17.11",
     "prop-types": ">=15.5.4",
     "react": ">=16.8.6",
     "react-dom": ">=16.8.6",
-    "styled-system": ">=4.1.0"
+    "styled-system": ">=5.0.18"
   },
   "devDependencies": {
     "@emotion/core": "^10.0.10",
@@ -64,13 +64,13 @@
     "@types/lodash": "^4.14.135",
     "@types/react": "^16.8.19",
     "@types/react-dom": "^16.8.4",
-    "@types/styled-system": "^4.1.0",
+    "@types/styled-system": "^5.0.1",
     "enzyme": "^3.9.0",
     "lodash": "^4.17.11",
     "prop-types": "^15.5.4",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "styled-system": "^4.1.0"
+    "styled-system": "^5.0.18"
   },
   "files": [
     "dist"
